Finalizacion de CRUD

+++ src/modules/daybook/views/EntryView.vue
@@ -8,14 +8,18 @@
       </div>
 
       <div>
-        <button class="btn btn-danger mx-2"
+        <input v-show="false" type="file" @change="onSelectedImage" ref="imageSelector" accept="image/png, image/jpg"/>
+
+        <button
+          class="btn btn-danger mx-2"
           v-if="entry.id"
-          @click="onDeleteEntry">
+          @click="onDeleteEntry"
+        >
           Borrar
           <i class="fa fa-trash-alt"></i>
         </button>
 
-        <button class="btn btn-primary">
+        <button class="btn btn-primary" @click="onSelectImage">
           Subir Foto
           <i class="fa fa-upload"></i>
         </button>
@@ -29,23 +33,29 @@
     </div>
 
     <img
-      src="https://c.tenor.com/Ny3gqWeXc38AAAAC/shiny-ditto-dance.gif"
+      v-if="entry.picture && !localImage"
+      :src="entry.picture"
+      alt="Entry Picture"
+      class="img-thumbnail"
+    />
+
+    <img
+      v-if="localImage"
+      :src="localImage"
       alt="Entry Picture"
       class="img-thumbnail"
     />
   </template>
 
-  <Fab 
-    icon="fa-save"
-    @on:click="saveEntry"
-  />
+  <Fab icon="fa-save" @on:click="saveEntry" />
 </template>
 
 <script>
 import { defineAsyncComponent } from "vue";
 import { mapGetters, mapActions } from "vuex";
 import getDayMonthYear from "../helpers/getDayMonthYear";
-import Swal from 'sweetalert2'
+import Swal from "sweetalert2";
+import uploadImage from '../helpers/uploadImage'
 
 export default {
   props: {
@@ -60,6 +70,8 @@ export default {
   data() {
     return {
       entry: null,
+      localImage: null,
+      file: null,
     };
   },
   computed: {
@@ -78,16 +90,18 @@ export default {
     },
   },
   methods: {
-    ...mapActions("journal", ["updateEntry","createEntry", "deleteEntry"]),
+    ...mapActions("journal", ["updateEntry", "createEntry", "deleteEntry"]),
     loadEntry() {
-      let entry
+      let entry;
 
-      if(this.id === 'new'){
+      this.localImage = null
+
+      if (this.id === "new") {
         entry = {
-          text: '',
-          date: new Date().getTime()
-        }
-      }else{
+          text: "",
+          date: new Date().getTime(),
+        };
+      } else {
         entry = this.getEntryById(this.id);
 
         if (!entry) return this.$router.push({ name: "no-entry" });
@@ -96,27 +110,65 @@ export default {
       this.entry = entry;
     },
     async saveEntry() {
-
       new Swal({
-        title: 'Espere por favor',
-        allowOutsideClick: false
-      })
-      Swal.showLoading()
+        title: "Espere por favor",
+        allowOutsideClick: false,
+      });
+      Swal.showLoading();
 
-      if(this.entry.id){
-        await this.updateEntry(this.entry)
-      }else{
-        const id = await this.createEntry(this.entry)
+      const picture = await uploadImage(this.file)
+      this.entry.picture = picture
+      console.log(picture)
 
-        this.$router.push({name: "entry", params: {id}})
+      if (this.entry.id) {
+        await this.updateEntry(this.entry);
+      } else {
+        const id = await this.createEntry(this.entry);
+
+        this.$router.push({ name: "entry", params: { id } });
       }
 
-      Swal.fire('Guardado', 'Entrada registrada exitosamente', 'success')
+      this.file = null
+
+      Swal.fire("Guardado", "Entrada registrada exitosamente", "success");
     },
-    async onDeleteEntry(){
-      await this.deleteEntry(this.entry.id)
-      this.$router.push({name: "no-entry"})
-    }
+    async onDeleteEntry() {
+      const { isConfirmed } = await Swal.fire({
+        title: "Estas seguro?",
+        text: "Una vez borrado no se puede recuperar",
+        showDenyButton: true,
+        confirmButtonText: "Si, estoy seguro",
+      });
+
+      if (isConfirmed) {
+        new Swal({
+          title: "Espere por favor",
+          allowOutsideClick: false,
+        });
+        Swal.showLoading();
+        await this.deleteEntry(this.entry.id);
+        this.$router.push({ name: "no-entry" });
+        Swal.fire("Eliminado", "", "success");
+      }
+    },
+    onSelectedImage(event) {
+      const file = event.target.files[0];
+
+      if (!file) {
+        this.localImage = null;
+        this.file = null;
+        return;
+      }
+
+      this.file = file;
+
+      const fr = new FileReader();
+      fr.onload = () => (this.localImage = fr.result);
+      fr.readAsDataURL(file);
+    },
+    onSelectImage() {
+      this.$refs.imageSelector.click();
+    },
   },
   created() {
     this.loadEntry();
@@ -142,7 +194,7 @@ textarea {
 }
 
 img {
-  width: 200px;
+  width: 150px;
   position: fixed;
   bottom: 150px;
   right: 20px;
